move _xtick_directional, _ytick_directional to plot module
